+++ cpp/api/lib/BUILD
@@ -1,21 +1,12 @@
 package(default_visibility = ["//visibility:public"])
 
-cc_binary(
-    name = "libtrtorch_plugins.so",
-    srcs = [],
-    deps = [
-        "//core/plugins:plugins"
-    ],
-    linkstatic = True,
-    linkshared = True
-)
 
 cc_binary(
     name = "libtrtorchrt.so",
     srcs = [],
     deps = [
         "//core/runtime:runtime",
-        "//core/plugins:plugins"
+        "//core/plugins"
     ],
     linkstatic = True,
     linkshared = True
